+++ exemple-files/attachments.js
@@ -11,6 +11,6 @@ module.exports =
     attachment:
     {
         type: "image",
-        url : "$assets_path/images/image.png"
+        url : "{$assets_path}/images/image.png"
     }
 }